@@ -1,22 +1,22 @@
 name: hasura_connect
 description: Connect your Flutter/Dart apps to Hasura simply. All the power of GraphQL and Postgres in Flutter.
-version: 2.0.1
+version: 3.0.0-nullsafety.0
 homepage: https://github.com/Flutterando/hasura_connect
 
 environment:
-  sdk: '>=2.8.1 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
-  async: ^2.4.2
-  dartz: ^0.9.1
-  http: '>=0.12.0 <1.0.0'
-  meta: '>=1.1.0 <1.3.1'
   websocket: ^0.0.5
-  string_validator: ^0.1.4
-  uuid: ^2.0.4
+  http: ^0.13.0-nullsafety.0
+  async: ^2.5.0-nullsafety.3
+  either_dart: ^0.1.0-nullsafety.1
+  meta: ^1.3.0-nullsafety.6
+  string_validator: ^0.2.0-nullsafety.0
+  uuid: ^3.0.0-nullsafety.0
+
 
 dev_dependencies:
   test_coverage: ^0.4.3
-  mockito: ^4.1.1
-  pedantic: ^1.9.0
-  test: ^1.14.4
+  mocktail: ^0.0.1-dev.9
+  pedantic: ^1.9.2
