@@ -4,66 +4,66 @@ import 'package:hasura_connect/src/domain/models/query.dart';
 import 'package:meta/meta.dart';
 
 class Snapshot<T> extends Stream<T> implements EventSink<T> {
-  StreamController controller;
-  Stream<T> rootStream;
-  Query query;
+  late StreamController _controller;
+  late Stream<T> _rootStream;
+  late Query _query;
+  Query get query => _query;
   final _WrapperStartWith<T> _wrapper = _WrapperStartWith<T>();
-  final void Function(Snapshot) closeConnection;
-  final void Function(Snapshot) changeVariablesF;
+  final void Function(Snapshot)? closeConnection;
+  final void Function(Snapshot)? changeVariablesF;
 
   Snapshot({
-    @required Query query,
+    required Query query,
+    Stream<T>? rootStream,
+    StreamController? controller,
     this.closeConnection,
-    this.rootStream,
-    this.controller,
-    T defaultValue,
+    T? defaultValue,
     this.changeVariablesF,
   }) {
     _wrapper.value = defaultValue;
-    this.query = query;
-    controller = controller ?? StreamController.broadcast();
-    rootStream = rootStream ??
-        controller.stream.transform(StartWithStreamTransformer<T>(_wrapper));
+    _query = query;
+    _controller = controller ?? StreamController.broadcast();
+    _rootStream = rootStream ?? _controller.stream.transform(StartWithStreamTransformer<T>(_wrapper));
   }
   Future changeVariables(Map<String, dynamic> variables) async {
-    query = query.copyWith(variables: variables);
-    changeVariablesF(this);
+    _query = _query.copyWith(variables: variables);
+    if (changeVariablesF != null) {
+      changeVariablesF!(this);
+    }
   }
 
   @override
-  StreamSubscription<T> listen(void Function(T event) onData,
-      {Function onError, void Function() onDone, bool cancelOnError}) {
-    return rootStream.listen(onData,
-        onError: onError, onDone: onDone, cancelOnError: cancelOnError);
+  StreamSubscription<T> listen(void Function(T event)? onData, {Function? onError, void Function()? onDone, bool? cancelOnError}) {
+    return _rootStream.listen(onData, onError: onError, onDone: onDone, cancelOnError: cancelOnError);
   }
 
   @override
   Snapshot<S> map<S>(S Function(T event) convert) {
     return Snapshot<S>(
-        query: query,
-        rootStream: rootStream.map((e) => convert(e)),
-        controller: controller,
+        query: _query,
+        rootStream: _rootStream.map((e) => convert(e)),
+        controller: _controller,
         closeConnection: closeConnection,
         changeVariablesF: changeVariablesF,
-        defaultValue: convert(_wrapper.value));
+        defaultValue: _wrapper.value == null ? null : convert(_wrapper.value!));
   }
 
   @override
   void add(T event) {
     _wrapper.value = event;
-    controller.add(event);
+    _controller.add(event);
   }
 
   @override
-  void addError(Object error, [StackTrace stackTrace]) {
-    controller.addError(error, stackTrace);
+  void addError(Object error, [StackTrace? stackTrace]) {
+    _controller.addError(error, stackTrace);
   }
 
   @mustCallSuper
   @override
   void close() {
-    controller.stream.drain();
-    controller.close();
+    _controller.stream.drain();
+    _controller.close();
     closeConnection?.call(this);
   }
 }
@@ -71,17 +71,15 @@ class Snapshot<T> extends Stream<T> implements EventSink<T> {
 class StartWithStreamTransformer<T> extends StreamTransformerBase<T, T> {
   final StreamTransformer<T, T> transformer;
 
-  StartWithStreamTransformer(_WrapperStartWith<T> wrapper)
-      : transformer = _buildTransformer<T>(wrapper);
+  StartWithStreamTransformer(_WrapperStartWith<T> wrapper) : transformer = _buildTransformer<T>(wrapper);
 
   @override
   Stream<T> bind(Stream<T> stream) => transformer.bind(stream);
 
-  static StreamTransformer<T, T> _buildTransformer<T>(
-      _WrapperStartWith wrapper) {
+  static StreamTransformer<T, T> _buildTransformer<T>(_WrapperStartWith wrapper) {
     return StreamTransformer<T, T>((Stream<T> input, bool cancelOnError) {
-      StreamController<T> controller;
-      StreamSubscription<T> subscription;
+      late StreamController<T> controller;
+      late StreamSubscription<T> subscription;
 
       controller = StreamController<T>(
           sync: true,
@@ -90,13 +88,9 @@ class StartWithStreamTransformer<T> extends StreamTransformerBase<T, T> {
               controller.add(wrapper.value);
             }
 
-            subscription = input.listen(controller.add,
-                onError: controller.addError,
-                onDone: controller.close,
-                cancelOnError: cancelOnError);
+            subscription = input.listen(controller.add, onError: controller.addError, onDone: controller.close, cancelOnError: cancelOnError);
           },
-          onPause: ([Future<dynamic> resumeSignal]) =>
-              subscription.pause(resumeSignal),
+          onPause: ([Future<dynamic>? resumeSignal]) => subscription.pause(resumeSignal),
           onResume: () => subscription.resume(),
           onCancel: () => subscription.cancel());
 
@@ -106,5 +100,5 @@ class StartWithStreamTransformer<T> extends StreamTransformerBase<T, T> {
 }
 
 class _WrapperStartWith<T> {
-  T value;
+  T? value;
 }
