@@ -6,9 +6,9 @@ import 'package:hasura_connect/src/presenter/hasura_connect_base.dart';
 import '../../domain/models/request.dart';
 
 abstract class Interceptor {
-  Future onRequest(Request request);
-  Future onResponse(Response data);
-  Future onError(HasuraError request);
+  Future onRequest(Request? request);
+  Future onResponse(Response? data);
+  Future onError(HasuraError? request);
   Future<void> onSubscription(Request request, Snapshot snapshot);
   Future<void> onConnected(HasuraConnect connect);
   Future<void> onTryAgain(HasuraConnect connect);
