@@ -1,4 +1,4 @@
-import 'package:dartz/dartz.dart';
+import 'package:either_dart/either.dart';
 
 import '../entities/response.dart';
 import '../errors/errors.dart';
@@ -6,5 +6,5 @@ import '../errors/errors.dart';
 import '../models/request.dart';
 
 abstract class RequestRepository {
-  Future<Either<HasuraError, Response>> sendRequest({Request request});
+  Future<Either<HasuraError, Response>> sendRequest({required Request request});
 }
