@@ -1,4 +1,4 @@
-import 'package:dartz/dartz.dart';
+import 'package:either_dart/either.dart';
 import 'package:string_validator/string_validator.dart';
 import '../repositories/request_repository.dart';
 import '../errors/errors.dart';
@@ -6,7 +6,7 @@ import '../models/request.dart';
 import '../entities/response.dart';
 
 abstract class QueryToServer {
-  Future<Either<HasuraError, Response>> call({Request request});
+  Future<Either<HasuraError, Response>> call({required Request request});
 }
 
 class QueryToServerImpl implements QueryToServer {
@@ -15,16 +15,15 @@ class QueryToServerImpl implements QueryToServer {
   QueryToServerImpl(this.repository);
 
   @override
-  Future<Either<HasuraError, Response>> call({Request request}) async {
+  Future<Either<HasuraError, Response>> call({required Request request}) async {
     if (!request.query.isValid) {
-      return Left(const InvalidRequestError('Invalid document'));
+      return Left(InvalidRequestError('Invalid document'));
     } else if (!request.query.document.startsWith('query')) {
-      return Left(const InvalidRequestError('Document is not a query'));
+      return Left(InvalidRequestError('Document is not a query'));
     } else if (request.type != RequestType.query) {
-      return Left(
-          const InvalidRequestError('Request type is not RequestType.query'));
+      return Left(InvalidRequestError('Request type is not RequestType.query'));
     } else if (!isURL(request.url)) {
-      return Left(const InvalidRequestError('Invalid url'));
+      return Left(InvalidRequestError('Invalid url'));
     }
 
     return await repository.sendRequest(request: request);
