@@ -1,6 +1,6 @@
 import 'package:hasura_connect/src/domain/entities/connector.dart';
 import 'package:hasura_connect/src/domain/errors/errors.dart';
-import 'package:dartz/dartz.dart';
+import 'package:either_dart/either.dart';
 import 'package:hasura_connect/src/domain/models/query.dart';
 import 'package:hasura_connect/src/domain/models/request.dart';
 import 'package:hasura_connect/src/domain/repositories/connector_repository.dart';
@@ -9,7 +9,7 @@ import 'package:hasura_connect/src/infra/datasources/connector_datasource.dart';
 class ConnectorRepositoryImpl implements ConnectorRepository {
   final ConnectorDatasource datasource;
 
-  ConnectorRepositoryImpl({this.datasource});
+  ConnectorRepositoryImpl({required this.datasource});
 
   @override
   Future<Either<HasuraError, Connector>> getConnector(String url) async {
