@@ -1,10 +1,10 @@
-import 'package:dartz/dartz.dart';
+import 'package:either_dart/either.dart';
 import 'package:hasura_connect/src/domain/entities/connector.dart';
 import 'package:hasura_connect/src/domain/errors/errors.dart';
 import 'package:hasura_connect/src/domain/repositories/connector_repository.dart';
 import 'package:hasura_connect/src/infra/datasources/connector_datasource.dart';
 import 'package:hasura_connect/src/infra/repositories/connector_repository.dart';
-import 'package:mockito/mockito.dart';
+import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
 
 class ConnectorDataourceMock extends Mock implements ConnectorDatasource {}
@@ -18,8 +18,7 @@ void main() {
   });
 
   test('should return Response', () async {
-    when(datasource.websocketConnector(''))
-        .thenAnswer((_) async => Connector(null));
+    when(datasource.websocketConnector('')).thenAnswer((_) async => Connector(null));
     final result = await repository.getConnector('');
     expect(result | null, isA<Connector>());
   });
@@ -31,8 +30,7 @@ void main() {
   });
 
   test('should return error from datasource', () async {
-    when(datasource.websocketConnector(''))
-        .thenThrow(InvalidRequestError('error'));
+    when(datasource.websocketConnector('')).thenThrow(InvalidRequestError('error'));
     final result = await repository.getConnector('');
     expect(result.fold(id, id), isA<InvalidRequestError>());
   });
