@@ -6,7 +6,7 @@ import 'package:hasura_connect/src/domain/errors/errors.dart';
 import 'package:hasura_connect/src/domain/models/query.dart';
 import 'package:hasura_connect/src/external/websocket_connector.dart';
 import 'package:hasura_connect/src/presenter/hasura_connect_base.dart';
-import 'package:mockito/mockito.dart';
+import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
 import 'package:http/http.dart' as http;
 import 'package:websocket/websocket.dart';
@@ -25,8 +25,7 @@ void main() {
   final wrapper = WrapperMock();
 
   setUp(() {
-    connect = HasuraConnect('https://fake-hasura.com',
-        interceptors: [LogInterceptor()]);
+    connect = HasuraConnect('https://fake-hasura.com', interceptors: [LogInterceptor()]);
     when(client.post(
       any,
       body: anyNamed('body'),
@@ -81,8 +80,7 @@ void main() {
       final snapshot = Snapshot(query: Query(document: 'null'));
       expect(snapshot, emits('test'));
       connect.snapmap['fdfhsf'] = snapshot;
-      connect.rootStreamListener(
-          {'id': 'fdfhsf', 'payload': 'test', 'type': 'data'});
+      connect.rootStreamListener({'id': 'fdfhsf', 'payload': 'test', 'type': 'data'});
       await snapshot.close();
     });
     test('should execute with HasuraRequestError', () async {
@@ -126,20 +124,12 @@ void main() {
     test('should execute connection_ack', () async {
       final snapshot = Snapshot(query: Query(document: 'null'));
       connect.snapmap['fdfhsf'] = snapshot;
-      final data = {
-        'id': 'fdfhsf',
-        'payload': 'test',
-        'type': 'connection_ack'
-      };
+      final data = {'id': 'fdfhsf', 'payload': 'test', 'type': 'connection_ack'};
       await connect.normalizeStreamValue(data);
       await snapshot.close();
     });
     test('should execute connection_error', () async {
-      final data = {
-        'id': 'fdfhsf',
-        'payload': 'test',
-        'type': 'connection_error'
-      };
+      final data = {'id': 'fdfhsf', 'payload': 'test', 'type': 'connection_error'};
       await connect.normalizeStreamValue(data);
     });
   });
